+++ src/routes/index.svelte
@@ -4,10 +4,10 @@
 	import type { Todo } from '$models/todo';
 
 	import Container from '$layouts/Container.svelte';
-	import Input from '$components/Input.svelte';
-	import InputGroup from '$components/InputGroup.svelte';
+	import Input from '$components/form/Input.svelte';
+	import InputGroup from '$components/form/InputGroup.svelte';
 	import NavBar from '$layouts/NavBar.svelte';
-	import SaveButton from '$components/SaveButton.svelte';
+	import SaveButton from '$components/form/SaveButton.svelte';
 	import Tab from '$components/tab/Tab.svelte';
 	import TabGroup from '$components/tab/TabGroup.svelte';
 	import TodoItem from '$components/todo/TodoItem.svelte';
